gmeet_api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
pymysql>=1.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

